+++ foundry/interfaces/ICallbackVerifier.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: UNLICENSED
-pragma solidity ^0.8.28;
+pragma solidity ^0.8.26;
 
 interface ICallbackVerifier {
     error UnauthorizedCaller(string exchange, address sender);
@@ -7,10 +7,8 @@ interface ICallbackVerifier {
     /**
      * @dev This method should revert if the sender is not a verified sender of the exchange.
      */
-    function verifyCallback(address sender, bytes calldata data)
-        external
-        returns (
-            uint256 amountOwed,
-            address tokenOwed
-        );
+    function verifyCallback(
+        address sender,
+        bytes calldata data
+    ) external returns (uint256 amountOwed, address tokenOwed);
 }